requirements.txt
pandas==1.2.4
seaborn==0.11.1
matplotlib==3.7.4
python-dateutil==2.8.2
requests==2.31.0
python-dotenv==1.0.0